test: consolidate CTO test suite, pin journal contract, fix CLI/TUI/docs parity drift
- Consolidated cto/ suite: linearTokenRefresh tests folded into linearAuth
  (helper-of-parent), linearLiveStatusService tests folded into
  linearLaneCardService (Linear PR-reflection contract); 6 → 4 test files,
  all 84 cases green.
- Added appendTurnJournal contract test (format, caps, blank-turn no-op).
- TUI: removed dead /linear workflows|run|route|sync|ingress slash groups that
  pointed at deleted daemon tools; usage strings updated; tests replaced.
- CLI: help text no longer advertises removed `ade cto` / linear workflow
  subcommands; stale routing comment fixed.
- Docs: restored Source file map to linear-integration README; fixed 3 stale
  ARCHITECTURE.md tree comments.
