+++ apps/ade-cli/src/tuiClient/commands.ts
@@ -91,17 +91,12 @@ export const BUILTIN_COMMANDS: BuiltinCommand[] = [
   { name: "/pr land", description: "Merge the active PR (needs confirm)", placement: "right", argumentHint: "[confirm] [merge|squash|rebase] [bypass]", category: "PRs" },
   { name: "/pr update-branch", description: "Update the PR branch with its base (merge or rebase)", placement: "right", argumentHint: "[merge|rebase]", category: "PRs" },
   { name: "/pr checks", description: "Show PR checks", placement: "right", category: "PRs" },
-  { name: "/linear", description: "Run Linear workflow, route, sync, or ingress commands", placement: "right", argumentHint: "<group>", category: "Linear" },
+  { name: "/linear", description: "Run Linear issue commands", placement: "right", argumentHint: "<group>", category: "Linear" },
   { name: "/linear list", description: "List Linear work", placement: "right", category: "Linear" },
-  { name: "/linear workflows", description: "List Linear workflow runs", placement: "right", category: "Linear" },
-  { name: "/linear run", description: "Inspect or resolve a Linear run", placement: "right", argumentHint: "<status|resolve|cancel|reroute>", category: "Linear" },
-  { name: "/linear route", description: "Route a Linear issue", placement: "right", argumentHint: "<cto|worker>", category: "Linear" },
-  { name: "/linear sync", description: "Operate Linear sync", placement: "right", argumentHint: "<dashboard|run|queue|resolve|detail>", category: "Linear" },
-  { name: "/linear ingress", description: "Inspect Linear ingress", placement: "right", argumentHint: "<status|events|webhook>", category: "Linear" },
   { name: "/linear pull", description: "Pull a Linear ticket into chat context", placement: "right", argumentHint: "<id>", category: "Linear" },
   { name: "/linear comment", description: "Comment on a Linear ticket", placement: "right", argumentHint: "<id> <text>", category: "Linear" },
   { name: "/linear comments", description: "Show comments on a Linear ticket", placement: "right", argumentHint: "<id>", category: "Linear" },
-  { name: "/linear status", description: "Show Linear sync status", placement: "right", category: "Linear" },
+  { name: "/linear status", description: "Show Linear status", placement: "right", category: "Linear" },
   { name: "/linear assign", description: "Assign a Linear ticket", placement: "right", argumentHint: "<id> <user>", category: "Linear" },
   { name: "/feedback", description: "Submit ADE feedback to GitHub issues", placement: "right", category: "Nav" },
   { name: "/chats", description: "List chats in the active lane", placement: "right", argumentHint: "[filter]", category: "Chats" },
